Add CameraInterface struct
This struct would hold an object that implements the new minimal
SerialConnection trait. This is done in order to make testing easier
by mocking the SerialConnection trait.

The new struct is not taken into use yet, the existing functions should
be cleaned up and usages be replaced with CameraInterface.

Also doing some refactoring to the Cargo.toml file

Additionally, still to be done:
- Adding debug logs for the new methods
- Refactor the different error types (std::io::Error vs anyhow)